:recycle: Re-arrange enum imports
