+++ platforms/platform-alexa/src/output/models/index.ts
@@ -1,3 +1,13 @@
+import { EnumLike } from '@jovotech/common';
+
+export enum ConfirmationStatus {
+  None = 'NONE',
+  Confirmed = 'CONFIRMED',
+  Denied = 'DENIED',
+}
+
+export type ConfirmationStatusLike = EnumLike<ConfirmationStatus>;
+
 export * from './AlexaOutputTemplate';
 export * from './AlexaOutputTemplateResponse';
 export * from './Directive';
